requirements.txt
# Core Framework
streamlit>=1.30.0

# AI/ML
google-generativeai>=0.3.0

# Environment Variables
python-dotenv>=1.0.0

# Data & Visualization
pandas>=2.0.0
plotly>=5.18.0

# HTTP Requests
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

